requirements.txt
black==22.1.0
cfgv==3.3.1
click==8.0.4
distlib==0.3.4
filelock==3.6.0
flake8==4.0.1
Flask==2.1.0
identify==2.4.12
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.22.3
opencv-python==4.5.5.64
pathspec==0.9.0
platformdirs==2.5.1
pre-commit==2.18.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
PyYAML==6.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
virtualenv==20.14.1
Werkzeug==2.1.0
